refactor:auth controller refactor auth 컨트롤러 부분을 수정했습니다.  #7
