+++ BE/src/routes/auth.ts
@@ -1,7 +1,5 @@
-import express, { Request, Response, NextFunction } from "express";
+import express from "express";
 import authController from "@controllers/auth";
-import UserModel from "@models/user";
-import { User } from "@interfaces/user";
 
 const router = express.Router();
 
@@ -14,16 +12,6 @@ router.get("/github/loginFail", authController.githubLoginFail);
 router.post("/apple", authController.apple);
 
 router.get("/logout", authController.logout);
-router.post("/register", async (req: Request, res: Response, next: NextFunction) => {
-  const user: User = {
-    id: null,
-    login_id: req.body.userID,
-    password: req.body.password,
-    img: "img1",
-    created_at: new Date(),
-  };
-  const result = await UserModel.insert(user, "USER");
-  res.json(result);
-});
+router.post("/register", authController.register);
 
 export = router;
